requirements.txt
# FastAPI and web server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
jinja2==3.1.6

# Note: torch, torchvision, torchaudio are installed separately in Dockerfile
# with specific CUDA 12.6 support for RTX 5070 Blackwell

# Diffusion models and AI
diffusers==0.30.3
transformers==4.47.0
accelerate==0.33.0
safetensors==0.4.4

# Image processing
pillow==12.0.0
opencv-python-headless==4.9.0.80
controlnet-aux==0.0.7
timm==0.9.16

# Scientific computing
numpy==1.26.3
scipy==1.12.0
einops==0.7.0

# NLP and tokenization
sentencepiece==0.2.0
protobuf==4.25.2

# Model hub
huggingface-hub==0.24.0

# Configuration
omegaconf==2.3.0

# External service clients
aiohttp==3.9.1
websockets==12.0

# Audio processing for InfiniteTalk
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
resampy==0.4.2

# Face processing for InfiniteTalk
mediapipe==0.10.9
face-alignment==1.4.1
imageio==2.33.0
imageio-ffmpeg==0.4.9
scikit-image==0.22.0

# InfiniteTalk specific
onnxruntime-gpu==1.16.3
insightface==0.7.3
# InfiniteTalk dependencies (will use on first generation)
kornia==0.7.0
einops-exts==0.0.4
rotary-embedding-torch==0.5.3
gitpython==3.1.40
